+++ web/pgadmin/utils/session.py
@@ -257,11 +257,10 @@ class FileBackedSessionManager(SessionManager):
         current_time = time.time()
         if not session.hmac_digest:
             session.sign(self.secret)
-        elif not session.force_write:
-            if session.last_write is not None and \
-                (current_time - float(session.last_write)) < \
-                    self.disk_write_delay:
-                return
+        elif not session.force_write and session.last_write is not None and \
+            (current_time - float(session.last_write)) < \
+                self.disk_write_delay:
+            return
 
         session.last_write = current_time
         session.force_write = False
@@ -402,6 +401,6 @@ def cleanup_session_files():
                     current_app.permanent_session_lifetime + \
                     datetime.timedelta(days=1)
 
-                if file_expiration_time <= datetime.datetime.now():
-                    if os.path.exists(absolute_file_name):
-                        os.unlink(absolute_file_name)
+                if file_expiration_time <= datetime.datetime.now() and \
+                        os.path.exists(absolute_file_name):
+                    os.unlink(absolute_file_name)
